broke peer location markers into a new component
